Added CUnit tests & Fixed some bugs
I reformatted all the code to our standard and fixed the bugs that I could. I created some files for testing, and edited the makefile to compile it, and also added some extra features to the makefile for convenience.

CUnit needs to be installed for testing. I put the command I used to install it in the README.txt file in src.